:tractor: :art: Refactor and move files around a bit
