+++ cmd/escuse-me/cmds/serve.go
@@ -1,0 +1,162 @@
+package cmds
+
+import (
+	"context"
+	"fmt"
+	"github.com/go-go-golems/glazed/pkg/cmds"
+	"github.com/go-go-golems/glazed/pkg/cmds/layers"
+	"github.com/go-go-golems/parka/pkg/glazed/handlers/datatables"
+	"github.com/go-go-golems/parka/pkg/handlers"
+	command_dir "github.com/go-go-golems/parka/pkg/handlers/command-dir"
+	"github.com/go-go-golems/parka/pkg/handlers/config"
+	template_dir "github.com/go-go-golems/parka/pkg/handlers/template-dir"
+	"github.com/go-go-golems/parka/pkg/server"
+	"github.com/pkg/errors"
+	"golang.org/x/sync/errgroup"
+	"os"
+	"os/signal"
+	"path/filepath"
+)
+
+type ServeCommand struct {
+	*cmds.CommandDescription
+	repositories []string
+}
+
+func (s *ServeCommand) runWithConfigFile(
+	ctx context.Context,
+	parsedLayers map[string]*layers.ParsedParameterLayer,
+	ps map[string]interface{},
+	configFilePath string,
+	serverOptions []server.ServerOption,
+) error {
+	return errors.New("not implemented")
+}
+
+func (s *ServeCommand) Run(
+	ctx context.Context,
+	parsedLayers map[string]*layers.ParsedParameterLayer,
+	ps map[string]interface{},
+) error {
+	// now set up parka server
+	port := ps["serve-port"].(int)
+	host := ps["serve-host"].(string)
+	debug := ps["debug"].(bool)
+	dev, _ := ps["dev"].(bool)
+
+	serverOptions := []server.ServerOption{
+		server.WithPort(uint16(port)),
+		server.WithAddress(host),
+		server.WithGzip(),
+	}
+
+	if configFilePath, ok := ps["config-file"]; ok {
+		return s.runWithConfigFile(ctx, parsedLayers, ps, configFilePath.(string), serverOptions)
+	}
+
+	configFile := &config.Config{
+		Routes: []*config.Route{
+			{
+				Path: "/",
+				CommandDirectory: &config.CommandDir{
+					Repositories: s.repositories,
+				},
+			},
+		},
+	}
+
+	contentDirs := ps["content-dirs"].([]string)
+
+	if len(contentDirs) > 1 {
+		return fmt.Errorf("only one content directory is supported at the moment")
+	}
+
+	if len(contentDirs) == 1 {
+		// resolve directory to absolute directory
+		dir, err := filepath.Abs(contentDirs[0])
+		if err != nil {
+			return err
+		}
+		configFile.Routes = append(configFile.Routes, &config.Route{
+			Path: "/",
+			TemplateDirectory: &config.TemplateDir{
+				LocalDirectory: dir,
+			},
+		})
+	}
+
+	server_, err := server.NewServer(serverOptions...)
+	if err != nil {
+		return err
+	}
+
+	if debug {
+		server_.RegisterDebugRoutes()
+	}
+
+	commandDirHandlerOptions := []command_dir.CommandDirHandlerOption{
+		command_dir.WithTemplateLookup(datatables.NewDataTablesLookupTemplate()),
+		command_dir.WithOverridesAndDefaultsOptions(
+		// ... override with ES server settings
+		),
+		command_dir.WithDefaultTemplateName("data-tables.tmpl.html"),
+		command_dir.WithDefaultIndexTemplateName(""),
+		command_dir.WithDevMode(dev),
+	}
+
+	templateDirHandlerOptions := []template_dir.TemplateDirHandlerOption{
+		template_dir.WithAlwaysReload(dev),
+	}
+
+	err = configFile.Initialize()
+	if err != nil {
+		return err
+	}
+
+	cfh := handlers.NewConfigFileHandler(
+		configFile,
+		handlers.WithAppendCommandDirHandlerOptions(commandDirHandlerOptions...),
+		handlers.WithAppendTemplateDirHandlerOptions(templateDirHandlerOptions...),
+		// TODO(manuel, 2023-12-13) This is the thing to implement
+		//handlers.WithRepositoryFactory()
+		handlers.WithDevMode(dev),
+	)
+
+	err = runConfigFileHandler(ctx, server_, cfh)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func runConfigFileHandler(
+	ctx context.Context,
+	server_ *server.Server,
+	cfh *handlers.ConfigFileHandler,
+) error {
+	err := cfh.Serve(server_)
+	if err != nil {
+		return err
+	}
+
+	ctx, cancel := context.WithCancel(ctx)
+	defer cancel()
+	ctx, stop := signal.NotifyContext(ctx, os.Interrupt)
+	defer stop()
+
+	errGroup, ctx := errgroup.WithContext(ctx)
+	errGroup.Go(func() error {
+		return cfh.Watch(ctx)
+	})
+	errGroup.Go(func() error {
+		return server_.Run(ctx)
+	})
+
+	err = errGroup.Wait()
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
