+++ pkg/cmd.go
@@ -7,10 +7,8 @@ import (
 	"github.com/elastic/go-elasticsearch/v8"
 	"github.com/elastic/go-elasticsearch/v8/esapi"
 	"github.com/go-go-golems/glazed/pkg/cmds"
-	"github.com/go-go-golems/glazed/pkg/cmds/alias"
 	"github.com/go-go-golems/glazed/pkg/cmds/layers"
 	"github.com/go-go-golems/glazed/pkg/cmds/layout"
-	"github.com/go-go-golems/glazed/pkg/cmds/loaders"
 	"github.com/go-go-golems/glazed/pkg/cmds/parameters"
 	"github.com/go-go-golems/glazed/pkg/helpers/files"
 	"github.com/go-go-golems/glazed/pkg/helpers/templating"
@@ -18,11 +16,7 @@ import (
 	"github.com/go-go-golems/glazed/pkg/settings"
 	"github.com/go-go-golems/glazed/pkg/types"
 	"github.com/pkg/errors"
-	"github.com/rs/zerolog/log"
-	"gopkg.in/yaml.v3"
 	"io"
-	"io/fs"
-	"path/filepath"
 	"strings"
 )
 
@@ -209,209 +203,4 @@ func (esc *ElasticSearchCommand) RunQueryIntoGlaze(
 	}
 
 	return nil
-}
-
-// ElasticSearchCommandLoader walks through a directory and finds all directories that end with
-// .escuse-me and loads the commands from there.
-// The layout of an .escuse-me directory is as follows:
-// - main.yaml (mandatory) contains the command description
-//   - contains flags, arguments, name, short
-//   - reference the query template file
-//
-// - the alias folder, which contains alias definitions in single yaml files
-// - the data folder, which contains additional data in json / csv / yaml format
-//   - this data is passed to the template at evaluation file,
-//     and can be used to store things like tags and constant strings, boost values and the like
-type ElasticSearchCommandLoader struct {
-	clientFactory ESClientFactory
-}
-
-func NewElasticSearchCommandLoader(
-	clientFactory ESClientFactory,
-) *ElasticSearchCommandLoader {
-	return &ElasticSearchCommandLoader{
-		clientFactory: clientFactory,
-	}
-}
-
-func (escl *ElasticSearchCommandLoader) LoadCommandAliasFromYAML(
-	s io.Reader,
-	options ...alias.Option,
-) ([]*alias.CommandAlias, error) {
-	return loaders.LoadCommandAliasFromYAML(s, options...)
-}
-
-func (escl *ElasticSearchCommandLoader) LoadCommandFromDir(
-	f fs.FS,
-	dir string,
-	options ...cmds.CommandDescriptionOption,
-) ([]cmds.Command, []*alias.CommandAlias, error) {
-	mainFilePath := filepath.Join(dir, "main.yaml")
-
-	parents := loaders.GetParentsFromDir(dir)
-	// strip last path element from parents
-	if len(parents) > 0 {
-		parents = parents[:len(parents)-1]
-	}
-
-	s, err := f.Open(mainFilePath)
-	// skip file does not exist
-	if err != nil {
-		if _, ok := err.(*fs.PathError); !ok {
-			return nil, nil, errors.Wrapf(err, "Could not open main.yaml file for command %s", dir)
-		}
-	}
-
-	if s == nil {
-		return []cmds.Command{}, []*alias.CommandAlias{}, nil
-	}
-
-	defer func(s fs.File) {
-		err := s.Close()
-		if err != nil {
-			log.Error().Err(err).Msg("Could not close file")
-		}
-	}(s)
-
-	escd := &EscuseMeCommandDescription{
-		Flags:     []*parameters.ParameterDefinition{},
-		Arguments: []*parameters.ParameterDefinition{},
-	}
-	err = yaml.NewDecoder(s).Decode(escd)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	// Add print query and explain parameters
-	escd.Flags = append(escd.Flags, &parameters.ParameterDefinition{
-		Name:     "print-query",
-		Help:     "Print the query that is generated",
-		Type:     "bool",
-		Required: false,
-	}, &parameters.ParameterDefinition{
-		Name:     "explain",
-		Help:     "Explain the query",
-		Type:     "bool",
-		Required: false,
-	})
-
-	queryTemplate := ""
-
-	//load query template, if present
-	if escd.QueryTemplate != "" {
-		queryTemplatePath := filepath.Join(dir, escd.QueryTemplate)
-		s, err := fs.ReadFile(f, queryTemplatePath)
-		if err != nil {
-			return nil, nil, err
-		}
-
-		queryTemplate = string(s)
-	} else {
-		return nil, nil, errors.New("No query template specified")
-	}
-
-	options_ := []cmds.CommandDescriptionOption{
-		cmds.WithName(escd.Name),
-		cmds.WithShort(escd.Short),
-		cmds.WithLong(escd.Long),
-		cmds.WithFlags(escd.Flags...),
-		cmds.WithArguments(escd.Arguments...),
-		cmds.WithParents(parents...),
-		cmds.WithLayout(&layout.Layout{
-			Sections: escd.Layout,
-		}),
-	}
-	options_ = append(options_, options...)
-
-	description := cmds.NewCommandDescription(
-		escd.Name,
-		options_...,
-	)
-
-	esc, err := NewElasticSearchCommand(description, escl.clientFactory, queryTemplate)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	aliases := []*alias.CommandAlias{}
-
-	// check for aliases in alias folder
-	aliasDir := filepath.Join(dir, "alias")
-	fi, err := fs.Stat(f, aliasDir)
-	if err != nil {
-		// skip file does not exist
-		if _, ok := err.(*fs.PathError); !ok {
-			return nil, nil, err
-		}
-	} else {
-		if fi.IsDir() {
-			entries, err := fs.ReadDir(f, aliasDir)
-			if err != nil {
-				return nil, nil, err
-			}
-			for _, entry := range entries {
-				// skip hidden files
-				if strings.HasPrefix(entry.Name(), ".") {
-					continue
-				}
-				fileName := filepath.Join(aliasDir, entry.Name())
-				if strings.HasSuffix(fileName, ".yaml") || strings.HasSuffix(fileName, ".yml") {
-					s, err := f.Open(fileName)
-					if err != nil {
-						return nil, nil, err
-					}
-					defer func(s fs.File) {
-						err := s.Close()
-						if err != nil {
-							log.Error().Err(err).Msg("Could not close file")
-						}
-					}(s)
-
-					aliases_, err := escl.LoadCommandAliasFromYAML(s)
-					if err != nil {
-						return nil, nil, err
-					}
-					for _, alias := range aliases_ {
-						alias.Source = fileName
-
-						alias.Parents = append(esc.Parents, esc.Name)
-						aliases = append(aliases, alias)
-					}
-				}
-			}
-		}
-	}
-
-	return []cmds.Command{esc}, aliases, nil
-}
-
-func (l *ElasticSearchCommandLoader) LoadCommandsFromFS(f fs.FS, dir string, options []cmds.CommandDescriptionOption, aliasOptions []alias.Option) ([]cmds.Command, []*alias.CommandAlias, error) {
-	var commands []cmds.Command
-	var aliases []*alias.CommandAlias
-
-	if strings.HasSuffix(dir, ".escuse-me") {
-		return l.LoadCommandFromDir(f, dir, options...)
-	}
-	entries, err := fs.ReadDir(f, dir)
-	if err != nil {
-		return nil, nil, err
-	}
-	for _, entry := range entries {
-		// skip hidden files
-		if strings.HasPrefix(entry.Name(), ".") {
-			continue
-		}
-		fileName := filepath.Join(dir, entry.Name())
-		if entry.IsDir() {
-			subCommands, subAliases, err := l.LoadCommandsFromFS(f, fileName, options, nil)
-			if err != nil {
-				return nil, nil, err
-			}
-			commands = append(commands, subCommands...)
-			aliases = append(aliases, subAliases...)
-		}
-	}
-
-	return commands, aliases, nil
-
 }
