+++ pkg/factory.go
@@ -1,0 +1,12 @@
+package pkg
+
+//func NewRepositoryFactory() handlers.RepositoryFactory {
+//	yamlFSLoader := loaders.NewYAMLFSCommandLoader(&ElasticSearchCommandLoader{})
+//	yamlLoader := &loaders.YAMLReaderCommandLoader{
+//		YAMLCommandLoader: &ElasticSearchCommandLoader{
+//			DBConnectionFactory: sql.OpenDatabaseFromDefaultSqlConnectionLayer,
+//		},
+//	}
+//
+//	return handlers.NewRepositoryFactoryFromReaderLoaders(yamlLoader, yamlFSLoader)
+//}
